refactor(project-structure): move and rename auth, config, and session layers to follow clean architecture
