+++ app/src/main/java/com/example/cue_practice_management_mobile/config/di/NetworkModule.kt
@@ -1,5 +1,8 @@
-package com.example.cue_practice_management_mobile.core.network
+package com.example.cue_practice_management_mobile.config.di
 
+import com.example.cue_practice_management_mobile.config.network.AuthInterceptor
+import com.example.cue_practice_management_mobile.core.data.api.AuthService
+import com.example.cue_practice_management_mobile.config.network.TokenAuthenticator
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -25,7 +28,7 @@ annotation class AuthRetrofit
 @InstallIn(SingletonComponent::class)
 object NetworkModule {
 
-    private const val BASE_URL = "http://192.168.20.21:3000/api/"
+    private const val BASE_URL = "http://10.0.2.2:3000/api/"
 
     @Provides
     @Singleton
