+++ app/src/main/java/com/example/cue_practice_management_mobile/domain/repositories/AuthRepository.kt
@@ -1,6 +1,6 @@
-package com.example.cue_practice_management_mobile.features.auth.repositories
+package com.example.cue_practice_management_mobile.domain.repositories
 
-import com.example.cue_practice_management_mobile.core.models.User
+import com.example.cue_practice_management_mobile.domain.models.User
 import com.example.cue_practice_management_mobile.features.auth.models.LoginRequest
 import com.example.cue_practice_management_mobile.features.auth.models.LoginResponse
 import com.example.cue_practice_management_mobile.features.auth.models.RefreshTokenResponse
